refactor(test): extract shared passthrough branch for test runners pytest and vitest had nearly identical 13-line passthrough branches. Extract run_passthrough into shared.rs with callbacks for arg preparation and command execution.
